PatchEmbeddingモジュールは汎用的なので抽出
